Enhance in_place construction (#66)
The current implementation required ad least a move-constructor for
in-place construction.
This made it impossible to directly, create an optional of a type
without copy and move operators.

A workaround is using the variadic emplace function, but it requires to
construct the object in two steps.

std::optional does not require any copy or move operation for in-place
construction